+++ resources/views/admin/pages/create.blade.php
@@ -55,6 +55,19 @@
                 </div>
 
                 <x-render name="taxonomy_fields" :params="['type' => 'page']" />
+
+                @if(setting('navigation.show_meta_fields'))
+                <div class="edit-box">
+                    <header>Metadados</header>
+                    <article>
+                        <x-meta-editor
+                            name="meta"
+                            :existingKeys="$existingMetaKeys"
+                            :values="$page->meta ?? []"
+                        />
+                    </article>
+                </div>
+                @endif
             </div>
 
             <div class="aside-column">